fix namespace issues
