--- app/bot/announcements/announcements_access.rb
@@ -1,23 +1,0 @@
-module Announcements
-  class AnnouncementsAccess
-
-    def initialize(bot)
-      @bot = bot
-      listener
-    end
-
-    def fetch num
-      announcements_channel = @bot.channel ENV["ANNOUNCEMENTS_DISCORD_CHANNEL_ID"]
-      announcements_channel.history(num)
-    end
-
-    private
-
-    def listener
-      @bot.message(in: ENV["ANNOUNCEMENTS_DISCORD_CHANNEL_ID"].to_i) do |event|
-        Rails.cache.write('new_announcement', true, expires_in: 10.days)
-      end
-    end
-
-  end
-end
