+++ app/bot/draffle_messages.rb
