+++ app/bot/trade_alert_messages.rb
